requirements.txt
customtkinter
pillow
db-sqlite3